+++ projects/mainio-forms/src/lib/store/reducers/form-lifecycle-reducers.ts
@@ -161,6 +161,5 @@ function getQuestions(
       };
     })
   ];
-  console.log("Toret", original, tobeSet, toRet);
   return toRet;
 }